Extract cover components and update file paths
