+++ client/src/components/cover/Cover.js
@@ -1,9 +1,9 @@
 import React, {Fragment, useEffect, useContext} from 'react';
 import LoginModal from "./LoginModal"
 import RegisterModal from "./RegisterModal"
-import AuthContext from "../context/authentication/AuthContext"
+import AuthContext from "../../context/authentication/AuthContext"
 import M from "materialize-css/dist/js/materialize.min.js";
-import Alert from "../components/Alert"
+import Alert from "../Alert"
 
 const Cover = (props) => {
     const authContext = useContext(AuthContext);
