+++ client/src/components/cover/RegisterModal.js
@@ -1,5 +1,5 @@
 import React, {useContext, useState} from 'react';
-import AuthContext from "../context/authentication/AuthContext";
+import AuthContext from "../../context/authentication/AuthContext";
 
 const RegisterModal = (props) => {
     const authContext = useContext(AuthContext);
